Move grid.grid to grid.polygon to make ground for generalization
