+++ src/instamatic/grid/polygon.py
